correction small mistake 2

+++ Backend/src/main/java/tn/esprit/kaddem/controller/EtudiantController.java
@@ -91,7 +91,7 @@ public class EtudiantController {
     }
 
     @GetMapping("/findByProjet/{idEtu}")
-    public List<Projet> findByProjet(@PathVariable("idEtu") int idEtudiant){
+    public List<Projet> findByProjet(@PathVariable("idEtu") Long idEtudiant){
         return etudiantServices.getProjets(idEtudiant);
     }
